requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
crispy-bootstrap5==0.6
cryptography==35.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.9
django-allauth==0.45.0
django-anymail==8.4
django-countries==7.2.1
django-crispy-forms==1.13.0
Django-stackoverflow==0.1.7
gitdb==4.0.9
GitPython==3.1.24
idna==2.5
lifter==0.4.1
oauthlib==3.1.1
persisting-theory==0.2.1
pycparser==2.21
PyGithub==1.55
pygithub3==0.5.1
PyJWT==2.3.0
PyNaCl==1.4.0
python3-openid==3.2.0
pytz==2021.3
requests==2.18.3
requests-oauthlib==1.3.0
six==1.16.0
smmap==5.0.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.22
wrapt==1.13.3
